\name{chooseCourse}
\alias{chooseCourse}
\title{Choose course to begin}
\usage{
  chooseCourse()
}
\value{
  list containing directory name for chosen course and the
  course name
}
\description{
  Prompts the user to choose the course they would like to
  take and returns the corresponding directory name and
  course name.
}

