\name{findTroubleTags}
\alias{findTroubleTags}
\title{Find trouble tags}
\usage{
  findTroubleTags(progressFilePath)
}
\arguments{
  \item{progressFilePath}{path to user progress file}
}
\value{
  tags list of tags associated with questions with which
  the user struggled
}
\description{
  Reviews user progress file for questions that the user
  struggled with and returns a list of tags associated with
  those questions
}

