\name{isYes}
\alias{isYes}
\title{Determines whether yes or no}
\usage{
  isYes(response)
}
\arguments{
  \item{response}{character vector containing response to a
  yes/no question}
}
\value{
  Logical, TRUE or FALSE
}
\description{
  Determines whether a response to a yes/no question is yes
  or no.
}

