\name{openingMenu}
\alias{openingMenu}
\title{Displays opening menu options}
\usage{
  openingMenu()
}
\value{
  List containing the starting module name, row number, and
  full paths of both the user info and user progress files,
  as well as a list of course directory name and course
  name. Note that the return value is a list with four
  elements, the third and fourth of which are lists with
  two elements each.
}
\description{
  Displays opening menu options for user and returns the
  user's starting location, including module name and row
  number.
}

