\name{progressMarkers}
\alias{progressMarkers}
\title{Prints progress markers}
\usage{
  progressMarkers(current.row, total.rows)
}
\arguments{
  \item{current.row}{numeric representing the current row
  that the user is on in the module spreadsheet}

  \item{total.rows}{numeric representing the total number
  of rows in the module spreadsheet}
}
\description{
  Generates visual progress markers in console.
}

