\name{recordIsCorrect}
\alias{recordIsCorrect}
\title{Records whether correct answer}
\usage{
  recordIsCorrect(is.correct, text.file.path)
}
\arguments{
  \item{is.correct}{logical, TRUE or FALSE}

  \item{text.file.path}{full file path of text file where
  string is recorded}
}
\description{
  Records TRUE or FALSE to designated text file. In the
  swirl function, it is used to denote whether user's
  answer is right or wrong.
}

