\name{recordString}
\alias{recordString}
\title{Records (appends) a string to designated text file}
\usage{
  recordString(my.string, text.file.path)
}
\arguments{
  \item{my.string}{character string}

  \item{text.file.path}{full file path of text file where
  string is recorded}
}
\description{
  In the swirl function, it is used to output progress to a
  progress file.
}

