\name{storeUserInfo}
\alias{storeUserInfo}
\title{Stores user info}
\usage{
  storeUserInfo()
}
\value{
  List containing the full file paths of the user info file
  and user progress file, respectively
}
\description{
  Stores first time user info in a new text file and
  returns the file path. Also creates a progress file and
  returns location of it.
}

