\name{swirl}
\alias{swirl}
\title{Runs the swirl program}
\usage{
  swirl()
}
\description{
  This function is the main component of the swirl
  interactive learning system. It gets all necessary
  information from the user and reads all module content
  directly from the location specified by the user. When
  the user quits the program via the Esc key, all progress
  is saved so he or she can resume their session at a later
  time.
}
\author{
  Nicholas A. Carchedi
}

