\name{chooseCourse}
\alias{chooseCourse}
\title{Choose course to begin}
\usage{
  chooseCourse(progress.file.path)
}
\value{
  courseDirName directory name for chosen course
}
\description{
  Prompts the user to choose the course they would like to
  take and returns the corresponding directory name.
}

