\name{openingMenu}
\alias{openingMenu}
\title{Displays opening menu options}
\usage{
  openingMenu()
}
\value{
  List containing the starting module name, row number, and
  full paths of both the user info and user progress files.
  Note that the return value is a list with three elements,
  the third of which is a list itself with two elements.
}
\description{
  Displays opening menu options for user and returns the
  user's starting location, including module name and row
  number.
}

