\name{swirl}
\alias{swirl}
\title{Runs the swirl program}
\usage{
  swirl()
}
\description{
  This function is the main component of the swirl
  interactive learning system. It gets all necessary
  information from the user to run the desired modules and
  record the user's progress as he/she works through those
  modules.
}
\author{
  Nicholas A. Carchedi
}

